Adds a spinner
