+++ client/js/_author/components/assessments/question_types/multiple_choice/add_option.jsx
