+++ client/js/_author/components/common/dot_loader.jsx
@@ -1,0 +1,11 @@
+import React      from 'react';
+
+export default function dotLoader(props) {
+  return (
+    <div className="loader">{props.children}</div>
+  );
+}
+
+dotLoader.propTypes = {
+  children: React.PropTypes.node,
+};
