+++ client/styles/_author/styles.scss
@@ -12,6 +12,7 @@
 @import "partials/multiple-choice";
 @import "partials/short-answer";
 @import "partials/file-upload";
+@import "partials/loader";
 // ...
 
 // Third-party
